+++ easypost/__init__.py
@@ -15,17 +15,12 @@ __version__ = VERSION
 version_info = VERSION_INFO
 SUPPORT_EMAIL = "support@easypost.com"
 
-
 # use urlfetch as request_lib on google app engine, otherwise use requests
 request_lib = None
-# use a max timeout equal to that of all customer-facing backend operations
-_max_timeout = 90
 try:
     from google.appengine.api import urlfetch
 
     request_lib = "urlfetch"
-    # use the GAE application-wide "deadline" (or its default) if it's less than our existing max timeout
-    _max_timeout = min(urlfetch.get_default_fetch_deadline() or 60, _max_timeout)
 except ImportError:
     try:
         import requests
@@ -61,8 +56,7 @@ except ImportError:
 # config
 api_key = None
 api_base = "https://api.easypost.com/v2"
-# use our default timeout, or our max timeout if that is less
-timeout = min(60, _max_timeout)
+timeout = 60
 
 
 USER_AGENT = "EasyPost/v2 PythonClient/{0}".format(VERSION)
@@ -249,9 +243,6 @@ class Requestor(object):
             "Authorization": "Bearer %s" % my_api_key,
             "Content-type": "application/json",
         }
-
-        if timeout > _max_timeout:
-            raise Error("`timeout` must not exceed %d; it is %d" % (_max_timeout, timeout))
 
         if request_lib == "urlfetch":
             http_body, http_status = self.urlfetch_request(method, abs_url, headers, params)